Refactor process preview logic into service
