+++ RuneFleet/MainForm.cs
@@ -1,15 +1,14 @@
 ﻿using RuneFleet.Interop;
 using RuneFleet.Models;
 using RuneFleet.Services;
-using System.Diagnostics;
 
 namespace RuneFleet
 {
     public partial class MainForm : Form
     {
         private List<Account> accounts = new();
-        // Thumbnails
-        private Dictionary<IntPtr, IntPtr> thumbnailMap = new();
+        // Service handling process thumbnails and client watching
+        private ProcessDisplayService processService;
         // Keybinds
         private const int WM_HOTKEY = 0x0312;
         // Unique IDs for each keybind
@@ -28,6 +27,12 @@ namespace RuneFleet
             accounts = AccountLoader.LoadFromCsv("accounts.csv");
             UpdateGroupView();
             UpdateListView("All");
+
+            processService = new ProcessDisplayService(
+                this,
+                flowPanelProcesses,
+                accounts,
+                () => UpdateListView(groupSelection.SelectedItem?.ToString() ?? "All"));
 
             // Handling the keybinds
             NativeMethods.RegisterHotKey(this.Handle, HOTKEY_ID_PGDN, 0, (uint)Keys.PageDown);
@@ -81,7 +86,7 @@ namespace RuneFleet
         // Refreshes the process display and updates the list view with the selected group.
         private void buttonLoadPreviews_Click(object sender, EventArgs e)
         {
-            RefreshProcessDisplay();
+            processService.RefreshProcessDisplay();
             UpdateListView(groupSelection.SelectedItem?.ToString() ?? "All");
         }
 
@@ -166,139 +171,6 @@ namespace RuneFleet
             base.WndProc(ref m);
         }
 
-        // TODO: Restructure this to a different folder
-        // Refactored: Extracted logic into helper methods, improved process validation, and clarified event handling.
-        // Refreshes the process display by checking running processes and updating thumbnails.
-        private void RefreshProcessDisplay()
-        {
-            CleanupThumbnailsAndControls();
-
-            foreach (var acc in accounts)
-            {
-                if (!acc.Pid.HasValue)
-                    continue;
-
-                Process? proc = null;
-                try
-                {
-                    proc = Process.GetProcessById(acc.Pid.Value);
-                    if (proc.HasExited || !IsOsrsClient(proc))
-                    {
-                        acc.Pid = null;
-                        continue;
-                    }
-                }
-                catch (Exception)
-                {
-                    acc.Pid = null;
-                    continue;
-                }
-
-                AddProcessThumbnail(acc, proc);
-            }
-        }
-
-        // Cleans up thumbnails and controls to prevent memory leaks and UI clutter.
-        // TODO: Refactor this to a different folder
-        private void CleanupThumbnailsAndControls()
-        {
-            foreach (var thumb in thumbnailMap.Values)
-            {
-                NativeMethods.DwmUnregisterThumbnail(thumb);
-            }
-            thumbnailMap.Clear();
-            flowPanelProcesses.Controls.Clear();
-        }
-
-        // Checks if the process is an OSRS client based on its name.
-        // TODO: Refactor this to a different folder
-        // This method can be extended to include other clients if needed.
-        // TODO: Use this in other functions that check for OSRS clients ex. WatchForClientsAsync.
-        private bool IsOsrsClient(Process proc)
-        {
-            return (proc.ProcessName.Contains("osclient", StringComparison.OrdinalIgnoreCase) ||
-                proc.ProcessName.Contains("runelite", StringComparison.OrdinalIgnoreCase));
-        }
-
-        // Adds a thumbnail for the process in the flow panel.
-        // TODO: Refactor this to a different folder
-        private void AddProcessThumbnail(Account acc, Process proc)
-        {
-            var hwnd = proc.MainWindowHandle;
-            if (hwnd == IntPtr.Zero)
-                return;
-
-            var pictureBox = CreateProcessPictureBox(acc, proc);
-
-            flowPanelProcesses.Controls.Add(pictureBox);
-
-            if (NativeMethods.DwmRegisterThumbnail(this.Handle, hwnd, out IntPtr thumb) == 0)
-            {
-                SetThumbnailProperties(pictureBox, thumb);
-                thumbnailMap[hwnd] = thumb;
-            }
-        }
-
-        // Creates a PictureBox for the process thumbnail with event handling for clicks.
-        // TODO: Refactor this to a different folder
-        private PictureBox CreateProcessPictureBox(Account acc, Process proc)
-        {
-            var pictureBox = new PictureBox
-            {
-                Width = 99,
-                Height = 65,
-                BackColor = Color.Black,
-                Margin = new Padding(1)
-            };
-
-            pictureBox.MouseClick += (s, e) =>
-            {
-                if (e.Button == MouseButtons.Left)
-                {
-                    ClientHelper.FocusWindowByPid(acc.Pid.Value);
-                }
-                else if (e.Button == MouseButtons.Right)
-                {
-                    try
-                    {
-                        proc.Kill();
-                    }
-                    catch
-                    {
-                        // Optionally log or handle process kill exceptions
-                    }
-                    acc.Pid = null;
-                    RefreshProcessDisplay();
-                    UpdateListView(groupSelection.SelectedItem?.ToString() ?? "All");
-                }
-            };
-
-            return pictureBox;
-        }
-
-        // Sets the properties for the thumbnail, including position and visibility.
-        // TODO: Refactor this to a different folder
-        private void SetThumbnailProperties(PictureBox pictureBox, IntPtr thumb)
-        {
-            Point screenPos = pictureBox.PointToScreen(Point.Empty);
-            Point formPos = this.PointToClient(screenPos);
-
-            var props = new DWM_THUMBNAIL_PROPERTIES
-            {
-                dwFlags = DwmFlags.DWM_TNP_RECTDESTINATION | DwmFlags.DWM_TNP_VISIBLE | DwmFlags.DWM_TNP_OPACITY,
-                fVisible = true,
-                opacity = 255,
-                rcDestination = new RECT
-                {
-                    Left = formPos.X,
-                    Top = formPos.Y,
-                    Right = formPos.X + pictureBox.Width,
-                    Bottom = formPos.Y + pictureBox.Height
-                }
-            };
-
-            NativeMethods.DwmUpdateThumbnailProperties(thumb, ref props);
-        }
 
         // Starts or stops watching for new character processes based on the button state.
         private void buttonWatchCharacters_Click(object sender, EventArgs e)
@@ -317,7 +189,7 @@ namespace RuneFleet
                 labelLoading.Visible = true;
                 buttonWatchCharacters.Text = "Stop Import Helper";
                 // Watching for new client processes
-                watchTask = Task.Run(() => WatchForClientsAsync(watchTokenSource.Token));
+                watchTask = Task.Run(() => processService.WatchForClientsAsync(watchTokenSource.Token));
             }
             else
             {
@@ -336,60 +208,6 @@ namespace RuneFleet
             }
         }
 
-        // Watches for new RuneLite or OSRS client processes and updates the accounts list.
-        // This method runs in a separate task and checks for new processes every 2 seconds.
-        // TODO: Refactor this to a different folder
-        private async Task WatchForClientsAsync(CancellationToken token)
-        {
-            var knownPids = Process.GetProcesses()
-                .Where(p => p.ProcessName.Equals("RuneLite", StringComparison.OrdinalIgnoreCase) ||
-                            p.ProcessName.Equals("osclient", StringComparison.OrdinalIgnoreCase))
-                .Select(p => p.Id)
-                .ToHashSet();
-
-            while (!token.IsCancellationRequested)
-            {
-                var newClients = Process.GetProcesses()
-                    .Where(p => (p.ProcessName.Equals("RuneLite", StringComparison.OrdinalIgnoreCase) ||
-                                 p.ProcessName.Equals("osclient", StringComparison.OrdinalIgnoreCase)) &&
-                                 !knownPids.Contains(p.Id))
-                    .ToList();
-
-                foreach (var proc in newClients)
-                {
-                    knownPids.Add(proc.Id);
-
-                    try
-                    {
-                        var env = EnvReader.ReadEnvironmentVariablesFromProcess(proc.Id);
-                        string sessionId = env.ContainsKey("JX_SESSION_ID") ? env["JX_SESSION_ID"] : null;
-                        string displayName = env.ContainsKey("JX_DISPLAY_NAME") ? env["JX_DISPLAY_NAME"] : null;
-                        string characterId = env.ContainsKey("JX_CHARACTER_ID") ? env["JX_CHARACTER_ID"] : null;
-                        string clientPath = proc.MainModule?.FileName ?? "";
-
-                        if (!string.IsNullOrWhiteSpace(displayName) && !accounts.Any(a => a.DisplayName == displayName))
-                        {
-                            var acc = new Account
-                            {
-                                SessionId = sessionId,
-                                DisplayName = displayName,
-                                CharacterId = characterId,
-                                AccessToken = "",
-                                RefreshToken = "",
-                                Client = clientPath,
-                                Group = [],
-                                Arguments = ""
-                            };
-
-                            accounts.Add(acc);
-                            File.AppendAllText("accounts.csv", $"{acc.AccessToken},{acc.RefreshToken},{acc.SessionId},{acc.DisplayName},{acc.CharacterId},Captured;,{acc.Client},{acc.Arguments}\r\n");
-                        }
-                    }
-                    catch { }
-                }
-                await Task.Delay(2000, token);
-            }
-        }
 
         private void checkTopMost_CheckedChanged(object sender, EventArgs e)
         {
